Split up wasm module
